requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
langchain-core
langchain-openai
python-dotenv
langchain-community
langchain-postgres
langchain-cohere
psycopg[binary]
tqdm